Pass PageWaitForTimeoutTests
related to #526
